refactor: Split edgee into parts (CLI and server)
