+++ Cargo.toml
@@ -63,69 +63,4 @@ wasmtime-wasi = "27.0.0"
 edgee-sdk = "1.2.0"
 edgee-path = "0.1.0"
 
-[package]
-name = "edgee"
-description = "The full-stack edge platform for your edge oriented applications"
-version.workspace = true
-authors.workspace = true
-license.workspace = true
-keywords.workspace = true
-repository.workspace = true
-homepage.workspace = true
-edition.workspace = true
-
-[dependencies]
-addr.workspace = true
-aes.workspace = true
-anyhow.workspace = true
-base64.workspace = true
-brotli.workspace = true
-bytes.workspace = true
-cbc = { workspace = true, features = ["std"] }
-chrono = { workspace = true, features = ["serde"] }
-clap = { workspace = true, features = ["derive", "env"] }
-cookie.workspace = true
-edgee-path.workspace = true
-edgee-sdk.workspace = true
-futures.workspace = true
-hex.workspace = true
-html-escape.workspace = true
-http.workspace = true
-http-body-util.workspace = true
-hyper = { workspace = true, features = ["full"] }
-hyper-rustls = { workspace = true, features = ["http2", "ring"] }
-hyper-util = { workspace = true, features = ["full"] }
-ipnetwork.workspace = true
-json_comments.workspace = true
-json_pretty.workspace = true
-lazy_static.workspace = true
-libflate.workspace = true
-log.workspace = true
-openssl.workspace = true
-pin-project.workspace = true
-rand.workspace = true
-regex.workspace = true
-reqwest.workspace = true
-rustls = { workspace = true, features = ["ring"] }
-rustls-pemfile.workspace = true
-rustls-pki-types.workspace = true
-serde = { workspace = true, features = ["derive"] }
-serde_json.workspace = true
-serde_with.workspace = true
-serde_yml.workspace = true
-tokio = { workspace = true, features = ["full"] }
-tokio-rustls = { workspace = true, features = ["ring"] }
-toml.workspace = true
-tower = { workspace = true, features = ["full"] }
-tower-http = { workspace = true, features = ["full"] }
-tracing.workspace = true
-tracing-subscriber = { workspace = true, features = ["env-filter", "json"] }
-url.workspace = true
-uuid = { workspace = true, features = ["v4", "serde"] }
-wasmtime.workspace = true
-wasmtime-wasi.workspace = true
-
-[features]
-bundled = [
-    "openssl/vendored",
-]
+edgee-server = { version = "0.5.0", path = "crates/server" }
