+++ crates/cli/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "edgee"
+description = "The full-stack edge platform for your edge oriented applications"
+version.workspace = true
+authors.workspace = true
+license.workspace = true
+keywords.workspace = true
+repository.workspace = true
+homepage.workspace = true
+edition.workspace = true
+
+[dependencies]
+clap = { workspace = true, features = ["derive", "env"] }
+serde_yml.workspace = true
+tracing.workspace = true
+tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
+toml.workspace = true
+tracing-subscriber = { workspace = true, features = ["env-filter", "json"] }
+
+edgee-server.workspace = true
