+++ crates/cli/src/logger.rs
@@ -26,9 +26,7 @@ pub fn init(log_format: LogFormat, log_filter: Option<String>) {
     use tracing_subscriber::prelude::*;
     use tracing_subscriber::{fmt, EnvFilter};
 
-    use crate::config;
-
-    let config = &config::get().log;
+    let config = &edgee_server::config::get().log;
 
     let with_target = log_filter.is_none();
     let fmt_layer = fmt::layer().with_target(with_target);
