+++ crates/server/Cargo.toml
@@ -7,3 +7,52 @@ license.workspace = true
 keywords.workspace = true
 repository.workspace = true
 homepage.workspace = true
+
+[dependencies]
+addr.workspace = true
+aes.workspace = true
+anyhow.workspace = true
+base64.workspace = true
+bytes.workspace = true
+cbc = { workspace = true, features = ["std"] }
+chrono = { workspace = true, features = ["serde"] }
+cookie.workspace = true
+futures.workspace = true
+hex.workspace = true
+html-escape.workspace = true
+http-body-util.workspace = true
+http.workspace = true
+hyper = { workspace = true, features = ["server"] }
+hyper-rustls = { workspace = true, features = ["ring", "http2"] }
+hyper-util = { workspace = true, features = ["server", "service", "tokio", "http1", "http2"] }
+ipnetwork.workspace = true
+json_comments.workspace = true
+json_pretty.workspace = true
+lazy_static.workspace = true
+openssl.workspace = true
+pin-project.workspace = true
+regex.workspace = true
+reqwest.workspace = true
+rustls = { workspace = true, features = ["ring"] }
+rustls-pemfile.workspace = true
+rustls-pki-types.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json.workspace = true
+serde_with.workspace = true
+tokio = { workspace = true, features = ["rt", "net"] }
+tokio-rustls.workspace = true
+tower = { workspace = true, features = ["util"] }
+tower-http = { workspace = true, features = ["compression-full", "decompression-full"] }
+tracing.workspace = true
+url.workspace = true
+uuid = { workspace = true, features = ["v4", "serde"] }
+wasmtime.workspace = true
+wasmtime-wasi.workspace = true
+
+edgee-path.workspace = true
+edgee-sdk.workspace = true
+
+[features]
+bundled = [
+    "openssl/vendored",
+]
