+++ crates/server/src/proxy/compute/html.rs
