+++ crates/server/src/tools/crypto.rs
