+++ crates/server/wit/protocols.wit
